Move merging tests to `test_snps.py`
